+++ tests/fnet.rs
@@ -94,7 +94,7 @@ fn fnet_for_sequence_classification() -> anyhow::Result<()> {
 
     let sentiment_config = SentimentConfig {
         model_type: ModelType::FNet,
-        model_resource: model_resource,
+        model_resource,
         config_resource,
         vocab_resource,
         ..Default::default()